backend/requirements.txt
six==1.16.0
reportlab==3.5.68
django_filter==2.4.0
Django==3.0.5
djangorestframework==3.12.4
djoser==2.1.0
